Use $wgVisualEditorPluginModules to register VE listener
Plugin modules always run when VE is loaded, whereas
the activation hook only fires when VE has successfully initialized,
which means we miss events fired from code paths that don't lead to
successful activation.

Using the plugin system also means that we don't have to check whether
VE is present (either in the JS or when loading the module), because
the plugin system handles that for us.

However, this does mean that WikimediaEvents must be loaded after VE.

Change-Id: I7e059e759b722a1723b77f5d9b14afbf116d04c1