Log usage of deprecated JavaScript functionality
See https://meta.wikimedia.org/wiki/Schema:DeprecatedUsage

This will let us measure the decay in usage of functions that we
deprecate, and it will let us project the number of users that would be
impacted by the removal of a deprecated function, so that we can tune our
deprecation schedule to the actual speed at which wikis and extensions
adapt.
No information about the user is included in the schema; it is simply
measuring the rate at which cruft disappears.

Also deletes unused moduleStorage module.

Change-Id: Ieac7fead764b2a0d8295ff42fbcb345aa42e119c