Add experiment for testing sendBeacon reliability
The goals of the experiment are to get emperical data on
what browsers support sendBeacon, what proportion of our users
have these browsers, and the reliability of sendBeacon.

For browsers that do support it, we log the same event twice
(except noting which event is which) with the two methods
to allow correlating which events are not delivered.

For example, it will show what proportion of img events
come from a sendBeacon-supporting browser, but have no accompanying
sendBeacon event (or vice-versa).

For browsers without sendBeacon, there is nothing to correlate,
so if an img event is not delivered, we have no way to know.

As an optimization, only 1/10,000 views do the logging.

To make future data-gathering like this roll on and off faster,
set up a convention of a 'ext.wikimediaEvents' module that is always
present for this purpose (with or without code).  This solves
ResourceLoader issues (mainly related to cached anon HTML).

Bug: T44815
Change-Id: I07ecbde39b5f1ce640f40430b198f8c160a89d13