Real-time active editor metrics
Whenever a non-bot user edits a mainspace page, enqueue a deferred update which
checks if the edit is the fifth such edit by this user this month. If it is,
then the user now counts as an "active editor", per the metric defined at
https://www.mediawiki.org/wiki/Analytics/Metric_definitions#Active_editor.
Log an event and increment a stat.

Depends on change I8c0a93713c1 in mediawiki/core.

Change-Id: I370589c77059a4ae5b8e607285cd95587e43d6b7