Log backend response time on edit requests
Use a PageContentSaveComplete hook handler (to make sure that we cover all the
code-paths that result in an edit) to schedule a DeferredUpdate (to make sure
we run after the response has been flushed) to measure and report the time from
request initialization to the output being flushed to the user. Record it as
'editResponseTime'. Follows I36b29162ffcc in core.

Change-Id: Ic99ac31f740956aa7f8f84efabc88045ebdb69ef