Add "user_wpzero" AbuseFilter variable
This adds a variable to the AbuseFilter system allowing administrators
to filter based on whether an edit was made using a connection from a
Wikipedia Zero carrier or not. It checks for the presence of a
'X-Carrier' header, which is set by Varnish.

Any filter hit will log whether the editor was editing over a Wikipedia
Zero carrier, but not any details of specific carriers. Conceptually
this is similar to the "tor_exit_node" variable added by the TorBlock
extension.

Bug: T131211
Change-Id: I6e307866e00ef56fcd02af24fc2527e950bf5132