Add textcat subtest
Increases sampling from 0.5% to 1%. The additional 0.5% of users are
split into three buckets: a, b and c. These buckets provide the
cirrusUserTesting parameter to the backend which will set all the
appropriate configuration for their searches to be used as part of the
textcat A/B/C test.

This depends on user testing triggers in mediawiki-config from
Id78fb0de06. The config change should be deployed first.

This also depends on cirrus outputing extra textcat specific data in
I44fded9772.

Bug: T121542
Change-Id: Ia855716b930d7c83517687cc7c196df161dfd9b9