Turn off textcat subtest of search satisfaction
This test has been completed, so is safe to turn off. This is a partial
revert of Ia855716b9. By only partialy reverting it should be easier to
add back new tests in the future with less review.  Two jshint overrides
had to be added for things that are currently unused but will be used
again in the near future.

Bug: T134319
Change-Id: I6542ff7b35a64aef9bc2d6cbf20e56e187697cd6