search satisfaction: Check session is still active after page load
We were only checking if the session was still active when loading the
page. This opens the possibility for users to leave a page open for some
time and keep triggering events (perhaps by opening links in a new tab)
after their session has timed out. This checks session state on each
event rather than only once on page load.

Change-Id: I8155a4c2a985ff03c307b9cd19f28615173298d1