Take page visibility into account for search checkins
The page visibility api should let us get a slightly better look at how
long users are actually looking at the page. caniuse.com reports this
is supported in >90% of browsers wordwide.

Page visibility isn't a 100% match to what we need. To accommodate
accessibility tools the page visibility spec allows for browsers to 
consider being obscured by other apps to still be "visible", but this
is still an improvement from where we are today. Specifically background
tabs, or pages visible on a locked machine, are no longer considered 
active for the purpose of checkins.

Bug: T145102
Change-Id: Ib093f0cb411a4ef45d4996c966c90bf1691497b3