visibilitychange: Consider all changes to imply hidden
Previously we only emitted "hidden" if the visibilitychange event
was emitted to change state from visible to hidden, not if it
changed from visible to hidden.

It shouldn't make any difference since we already check the initial
value as well. To verify that this doesn't matter, this commit
simplifies the code by removing the document.hidden check from
the change event handler. See also I8e8ec96fc.

Change-Id: Ibd0935bef8f76a0e1a675db5242e2d4e301b0233