Introduce the reading depth schema
A reading depth schema is enabled on all pages and logs at
a sampling rate (as defined in wgWMEReadingDepthSamplingRate)
events on page load and page unload for all browsers which
are beacon capable.

In a follow up patch it will take into account the visibility time
but for now visibleLength will always be the same as totalLength.

A follow up patch will also track the time of first paint when
available to give a more fair representation of session.

Note times are rounded up to closest integer given milliseconds
is more than enough precision.

Bug: T155639
Change-Id: Ifb6ceb51bbb949ad17da17a3b38e9d04c3bf6457