Hygiene: Simplify getFirstPaintTime dependencies
This fixes an inconsistency between the treatment of the IE and Chrome
in getFirstPaintTime: [window.]performance.timing.navigationTiming is
required for the former but not for the latter - for which we use the
value provided by Chrome.

Given that we access performance.timing in the body of the function,
drop the parameter and get performance.timing.navigationTiming directly.

Change-Id: I04928980f72d86db212e26ed698060252c0780b0