geoFeatures,kartographer: Move 10min token from cookie to sessionStorage
Given this is only meant to live for 10 minutes, move it to session
storage instead of sending it over the network with all page loads
on all wikis.

For now, keep the token shared between these two schemas since
that should be a separate change (if separated).

The old cookie will be unused and already had an expiry.

Bug: T110353
Change-Id: I772b08522017967090a23b3c66791d4f0cb20895