Performance: Move 'SearchSatisfaction' to only run on non-Minerva skins
This is a pretty huge amount of code, and it's being blindly downloaded
by all users on desktop and mobile, only to immediately stop executing
if the skin is Minerva. Instead, move the code to be a default skinScript,
which is then not loaded on Minerva.

Before this change ext.wikimediaEvents has an uncompressed size of
21.8 KiB as the second largest module loaded on every mobile page;
afterwards, it's cut by more than half to 10.7 KiB on mobile, whilst
desktop is unaffected.

Change-Id: I60a3a78fbbdf6ae00c9050a644b105dcc341c5ff