RCFilters: Log click actions on RC page and on pages linked from there
This logs clicks on the diff, history, page, user page, user talk,
user contribs, user block and rollback links on Special:Recentchanges,
and also clicks on those links (as well as edit, undo, thank and patrol)
on pages reached by clicking a link on the RC page.

The latter is done by adding ?fromrc=1 to the URL of the link
as it is clicked.

Bug: T158458
Change-Id: I1a79088f344908758e00c97599cdeebf4ad1789a