Add tracking of WMDE Spring 2017 campaign
This could just go in mediawiki-config but it is safer living here.

This code will ONLY run on dewiki.

WMDE has created a CentralNotice banner that will link to a landing
page somewhere on dewiki with campaign=wmde2017spring in the request.

The Campaigns extension will only track campaigns if the landing page
is the CreateAccount page, so manually catch anon users that have the
campaign & set a cookie.

When the user eventually gets to the CreateAccount page inject the
campaign back into the request (which the Campaigns extension will
then pickup)

Post registration look for the cookie again, and a view of a main
namespace page and trigger the GuidedTour while removing the cookie.

The campaign is planned to start on the 12th April 2017
The campaign is planned to end on the 22nd April 2017

Change-Id: Iba34607fd057154cc028b6052ea623ad53c71352