RecentChangesClicks: Don't modify URL in click handler
This causes browsers' "visited link" features to not work properly:
the URL that is navigated to when the user clicks the link is
different from the href of the link until they click, so the
user will never see these links in the visited color.
This impedes the RC patrolling experience for many users.

Commenting out this code for now, which breaks part of the
click tracking; if we want to unbreak that, we'll have to
find a way to add ?fromrc=1 back on the server side.

Bug: T158458
Change-Id: I60bb5ab84016b7a66e1d23e049cc5980a105fff3