MVP of human graded search relevance on article pages
This is an MVP for asking our readers about the relevance of an article
page to a curated list of queries. For the MVP (and only for the MVP)
the list of articles and queries is hard-coded into the javascript
itself. This allows a small scale evaluation to see if the data will be
useful, or simply a bunch of noise.

Bug: T171740
Change-Id: I173a46da75d6632305883a81adb932f2a95227c6