Encode human search relevance survey into html output
Moves the search relevance configuration from javascript into php
configuration so it can be run on a larger number of pages without
exploding the size of javascript delivered to everyone with all
that config.

The downside of this approach is that because the config is cached
into the page html any updates have to work their way through varnish
caching. For the initial test this should be fine as we can purge the
relevant pages over some small time period. In the future when we are
doing larger scale collection we just have to keep in mind that
configuration changes take around 7 days to propogate.

Bug: T174387
Change-Id: I7ae522155e67610d25b5857d7b3918559bce8bc7