Limit logged skins for print event only to vector and minerva
We want to keep print events for longer than 90 days so we can do
more complex analysis in the future. While we defined the purge
strategy we found out that skin information potentially might
identify users who're printing.

Because of that we decided to store only vector and minerva skin
information as those are not identifying users, we will use it as
a lose identification of mobile vs non-mobile users. For all other
skins we will store 'other'. It will help us identify the percentage
ratio between most popular skins (minerva & vector) and all other
skins.

Bug: T169730
Bug: T175395
Change-Id: If527d2085211e9d23562965943712fda5248d72a