Allow extensions to explicitly opt into the ReadingDepthSchema
As we run more A/B tests we'll want to ensure those A/B tests remain
synced with the ReadingDepthSchema which provides vital information
to those tests.

Rather than add complex code to each A/B test to ensure the bucketing
is correct I propose we allow expansion of the ReadingDepthSchema by
the A/B test itself via a `wikimedia.event.ReadingDepthSchema.enable`
hook

Note, since the ReadingDepthSchema is not an A/B test itself, it is
perfectly acceptable to expand the reach of the schema, provided we
are careful in ensuring that we use sensible sampling rates for any
A/B tests that may be running and adjust the ReadingDepthSchema sample
rate accordingly.

Bug: T193584
Change-Id: I18bd5519fee5a3253a3c8b7802452e0dc68fec60