ReadingDepth test should accept a boolean field for sample groups.
ReadingDepth should set a field that describes the bucket and test that
the readingDepth schema was triggered from. This is so that when readingDepth
is triggered from an external test, the data for that external test
won't pollute the data from the default sample group, which is called
'default_sample'.

This patch takes the approach of dynamically enabling the ReadingDepth
schema as module dependency when a client requests the functionality via
mw.track().

Bug: T191532
Change-Id: I6a2e7c1cfce61260c208cbd0e99550e540924b2f