Ensure Minerva has initialised before loading and executing ReadingDepth
In Minerva,  skins.minerva.scripts should be loaded BEFORE ReadingDepth
is enabled and the pageLoaded event sent.

To solve this, we considered
1) Loading this code explicitly in Minerva (however this has a high
performance cost delaying all events in WikimediaEvents)
2) Exporting a public function enableReadingDepth and adding a new module
for all non-Minerva skins to enable it (however this would have meant
lots of conditional code inside Minerva and WikimediaEvents)
3) This patch

This patch, although imperfect seems the cleanest way to do this.
It keeps the code contained inside WikimediaEvents and requires the
least amount of conditional code.

Bug: T204144
Change-Id: I328449f9786f86db983330b80b16633b70605669