Track backend search token for autocomplete results
We have a searchToken to associate backend logs with search
satisfaction for full text, but not autocomplete. It will
be useful to track this so we can determine autocomplete
examination probabilities and improve our ability to perform
offline evaluation of autocomplete.

The dependant patches are not strictly required, everything should
still run correctly without them, but they are necessary to make
the search token available to tracking.

Change-Id: I04e9316d84d44b7d23fc3e4ceb7dbafe5e6e4025
Depends-On: I1663906e2fd71f7df215e563b09a0b4fb8948ab8
Depends-On: Ibc4bdb16db4c4be2817d4c888bd74d05adf12c92
Bug: T205348