EditorJourney: Store hash salt in Redis with 24 hour TTL
To reduce exposure from a data breach, store the salt used for hashing URLs in
Redis with a 24 hour TTL. The key to look up the salt is based on hashing the
concatenated user ID and registration timestamp. The salt is generated with
PHP's random_bytes(). The HMAC algorithm defaults to 'whirlpool' instead of md5.

The salt is stored on user account creation. On subsequent page views, if the
salt cannot be retrieved from Redis, we will generate a new one. A log message
is written so we can see how often this happens although we need to enable the
logging channel for WikimediaEvents.

Bug: T208233
Change-Id: I6c7ac1e11532cf35a791cded1746ac40fe7ac6f3