Allow VE edits to be tagged based on '&campaign=...' URL parameter
Valid values can be configured like this:
    $wgWMEEditCampaigns = [ 'foo', ... ];
If the value is valid, a tag like 'campaign-foo' is added to the edit.

The parameter is passed to the action=visualeditoredit API request.
That API module then passes all unrecognized parameters to an internal
action=edit API request. After that module saves the page, we can
access the value of the parameter using the 'RecentChange_save' hook.

This works in VisualEditor (and NWE) on desktop and mobile (after
Ie34139cb68f90f34eb243f1bb964ef578e90dfb2 is merged). It does not
work in classic wikitext editor or in mobile wikitext editor.

Bug: T209132
Change-Id: I872ad9a7020562a6d7bafd14c522745d4cf7bc7a