Relax flaky test: don't assert the exact value of user_id
It's unclear how this is meant to work. Certainly nobody can be
bothered to set $this->tablesUsed correctly. Locally I found 3 tests in
AbuseFilter that cause this to fail, and another that derives from
ApiTestCase (ApiMobileViewConvertTitleTest). Pretty much every test that
derives from ApiTestCase will cause this to fail due to ApiTestCase::setUp()
creating a user without setting $this->tablesUsed.

Change-Id: I4f82411833c91ac850d8725c3d3ad3153806c2af