Allow directing a sample of users to PHP 7
During the transition to PHP 7, we want to be able to have a sample of
users use PHP 7-rendered pages consistently. This patch allows to define
a sample size and direct a growing percentage of users to PHP7.

While the beta feature is enabled, we won't send any logged-in user to
PHP 7, as we would interfere with the beta feature.

Bug: T216676
Change-Id: Id4c88f9f7c5c874d4cacafb272f2c75459b5496f
(cherry picked from commit 11e2c4a12d8080af623470487ec37bf58d046de3)