EditorJourney: Redact title / create params, drop page_title
The page title from output is problematic to obfuscate properly, and it's not
needed for analysis, so stop sending it. We want to know if users are
experiencing permission errors on view, so add a workaround to include those
events in the permission errors property (we otherwise only care about
permission errors on edit).

Also redact the title and create properties on the query, as they are also
problematic to obfuscate correctly.

Bug: T213974
Change-Id: Ic80591065502735c1b8536f7925114aefb9211ce