Increase search satisfaction logging by 10x
Across the board increase in data collection by 10x, up to a cap of
1:10 sampling. The end goal here is to give all wikis the same
sampling rate for data collection. The difficulty with the current
scheme has been that having different wikis record at different
sampling rates is that nothing is directly comparable between wikis
during exploratory analysis. Any analysis has to carefully take into
account what the sampling rate was at the time data was collected.

This is a first step, after we verify that the increase in volume
doesn't cause issues we can submit subsequent patches that bring
the remaining sampling rates in line at 1:10.

Bug: T197129
Change-Id: I18c7645cd6b0c5e543562abab3169ab3c2c22129