Give all wikis the same search schema sampling rate
Unify all the wikis at 0.125 (1 in 8) sampling rate. This rate
will likely still need to climb in the future to ensure we have
enough data to derive meaningful metrics about lower volume wikis,
but after several bumps all wikis now share the same rate.

While here simplify the oneIn method which operated two different
ways depending on if the variable provided was > 1. Simplify to
the singular % based parameter. This was convenient when we had a
variety sampling rates, but a single % is easy to interpret and set
for most purposes.

Bug: T197129
Change-Id: Id095708301060ce424219651b65dfae4c7724973