Temporarily log search to two schemas
We had the split between SearchSatisfaction and TestSearchSatisfaction2
schemas in the past because one went to hadoop and the other to mysql.
We no longer need the events in mysql and can process everything
out of hadoop, so start logging everything to SearchSatisfaction.

A followup patch deployed about a week after this one will stop
logging to TestSearchSatisfaction2 completely.

Change-Id: I3a72a11d21a3ad923f5b86e96fd35637b35c013d