Eventlogging requires numbers to be numbers
When pulling the sampling multiplier out of the session it may have
been coerced into a string. On the first page of a session events
work, but on subsequent pages when fetching from localStorage the
sampleMultiplier is a string.  Eventlogging doesn't like that and
requires us to really send a number when we say we will.

Bug: T229614
Change-Id: I008d81fc93651a6d5cf26d6e8b6434a8ed6fab84