Configure Test event stream to be sent via EventGate
This will allow us to mw.eventLog.logEvent() an event
conforming to the Test schema.  We don't need to add any instrumentation
to do this, we will just log it from the browser console.  We will
use this to test that the new Event Platform components all work
with legacy EventLogging schemas.

Bug: T238230
Change-Id: If1986b519430624af556519b24704ea5076ae95c