citationUsage: Reduce use of jQuery in a few places
Push the creation of the jQuery objects slightly deeper down the
stack so that more functions have a safely available HTMLElement
reference, thus not having uncertainty over whether prop() or
[0] might return undefined if the collection is empty.

In the places it opens up, adopt a few native DOM accesses where
this is simpler and with less call overhead (e.g. plain properties
and textContent getter).

Change-Id: I2aa8e9895c45d684c52fc60a7f1ecf6dd5627f61