SearchSatisfaction: Set fields for search widget A/B tests in Vector
Set the skin, skinVersion, and extraParams fields to appropriate values
in the low-level event logger as they are relevant for events sent while
the user is interacting with the search widget and afterwards.

Further, only set the skinVersion and extraParams fields when the user
is using the Vector skin.

Supporting changes:

- Use SearchSatisfaction@1.2.0, defined in I20309fa2

Bug: T256100
Depends-On: I20309fa23902ee8cc49a6f5c60d60188290c0659
Depends-On: I3fef99e5b94ca4dc221282f28c7608f84d8fb746
Change-Id: Ifb27c6c44d183d250b95fdffbd7974727956450f