SearchSatisfaction: Set isAnon field
This enables the analysis of logged in/out users separately, which is
desireable in the context of the search input location A/B test (see the
associated bug for detail).

Set the field to the appropriate value in the low-level event logger (as
in Ifb27c6c4), as it's relevant for all events set by the instrument.

Supporting changes:

- Use SearchSatisfaction@1.3.0, defined in Ifa4f0c93

Bug: T259250
Depends-On: Ifa4f0c93e74cc78f9836afc374941e3c187a47e7
Change-Id: I90eeffc17e416f8beda5233bb9020c8b7426dec4
(cherry picked from commit b26cde4df41621dab5024449a3d4ece2b80e86a2)