[searchSatisfaction] check validity of test buckets
Ensure that the subTest in the session is actually allowed in the
validBuckets. On the search result page verify that the backend has
run with the expected tests.
Propagate cirrusUserTesting param to pagination links so that it does
not get lost.
When an inconsistent is detected the events from this session will
be recorded in the 'mismatch' or 'invalid' subTest.

Bug: T262845
Depends-On: I4d9364545dab183b2c44fc570e458f5e136c8708
Change-Id: I93aa56e34e6c892403c09f2fd1ed579215801915