Move and merge mediawiki.errorLogger.js code
Because the mw.errorLogger API is loaded during the critical path we
want to minimize its size without rendering it unusable or unstable.

We achieve this by moving as much of the stack trace
parsing/normalization code to the ext.wikimediaEvents module, which
already did some processing of errors for filtering purposes, leaving
the mediawiki.base module to just define the API to bind to.

Bug: T249826
Change-Id: I49d3edceb702efda5c72288b8cc681f7b18cc9e6