clientError: Merge mediawiki.errorLogger.js code (again)
Because the mw.errorLogger API is loaded during the critical path we
want to minimize its size without rendering it unusable or unstable.

We achieve this by moving as much of the stack trace
parsing/normalization code to the ext.wikimediaEvents module, which
already did some processing of errors for filtering purposes, leaving
the mediawiki.base module to just define the API for developers to bind
to.

Bug: T249826
Change-Id: I01036bb8530a6eb2277383564baf0e0cbbeb3a78