Fix dynamically loaded instruments
As well as converting the ext.wikimediaEvents ResourceLoader module to
use packageFiles, I3906a1b1 also made the SearchSatisfaction,
DesktopWebUIActions, MobileWebUIActions, and InukaPageView instruments
load dynamically based on the current skin. This was achieved by
defining a callback,
WikimediaEvents\WikimediaEventsHooks::getModuleFile(), that gets the
path to the file in which the instrument is defined.

However, the callback was incorrectly defined as accepting the name of
the instrument as its second parameter and not its third - the Resource
Loader's config is passed as the second parameter [0] - and so always
returned the empty string. Consequently, the instruments mentioned above
weren't being loaded.

[0] https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/core/+/refs/heads/wmf/1.36.0-wmf.32/includes/resourceloader/ResourceLoaderFileModule.php#1338

Change-Id: Ief0e11ce7d313c3f935c45ad0a6521dbd7c82d73