Fix: Restore exporting wgWMESchemaEditAttemptStepSamplingRate to JS
More fallout from the recent change to config loading in WikimediaEvents
(https://gerrit.wikimedia.org/r/c/mediawiki/extensions/WikimediaEvents/+/553356).

A recent change stopped exporting wgWMESchemaEditAttemptStepSamplingRate
to the global JS config and confined it to a virtual config file for the
extension. Unfortunately, it was needed by multiple other components. This
restores it.

Change-Id: I77210d7fa2317603eb557708a50ceb928bc3e30c