Revise file url logging rules
Define shouldLogFileUrl code more readable.

Apply following changes:
* Do not log file uris prefixed with "C:\\", "javascript:", "jar:"
- these are generally
added by bookmarklets. We don't want
to log errors associated for the same reason as blob:
* Do not log URIs with "/inj_js/" in the path. This is common on our
deployed wikis and cannot be filtered since it is not a prefix or
suffix
* We were dropping various errors which have no file url
in Chrome due to loading code from localStorage. We now
are in a good state to log these.

Bug: T266517
Change-Id: I53b03dd5d56ad18ae9d12d7d60a068aeb41f2b73