Add UniversalLanguageSelector instrument
Per T275894, Readers Web intend to increase the scope of the
UniversalLanguageSelector instrument to help them understand how users
are interacting with the Languages list in the sidebar.

In order to keep the UniversalLanguageSelector extension focussed and
not overly Wikimedia-specific, it was agreed that the instrument would
be migrated to the WikimediaEvents extension, where other such
instruments are located.

Bug: T275894
Change-Id: I2acee8ba8dc59c497dc6ba42cd5bd88dfd985c8b