Add properties to UniversalLanguageSelector events
Readers Web will be A/B testing the existing language switcher treatment
against a new one being worked on as part of Desktop Improvements.

Adding the web_session_id, isAnon, and userEditBucket properties will
allow us to better understand baseline user behavior for both logged-in
and logged-out users, as well as logged-in power users.

Adding the selectedInterfaceLanguage property allows us to understand
which language the user selected when they switched.

Bug: T275766
Depends-On: Ifb86f62125bffa42f09243716c68d747a05fb1e8
Change-Id: If3349003367013f4b20cd5aa3d00568fb4cc84e1