searchSatisfaction: Handle edge case in referrer URLs
Some wikis might use a custom search implementation and/or deliver
gadgets to the user that modify the search form. In the case of
wikidatawiki, something is adding a hidden input named "search" to the
form, which results in referrer.query.search being an array of duplicate
strings rather than a string.

If possible, wikidatawiki's search implementatation will be fixed in a
follow-up patch/edit but, for now, handle the edge case by plucking the
first value from referrer.query.search when it's an array.

Bug: T276474
Change-Id: I45ddc754b4eda939fdaa68d7923cbf7613005918