onUserSaveOptions: Warn and abort when originalOptions are empty
Change I7033142c0a15d95e1eb209cd2eba87faf8543fd0 fixes a bug in
MobileFrontend that causes numerous spurious PrefUpdates to be
issued. This patch aims to uncover any remaining cases of similar
behavior, prevent regressions, and provide a measure of future-
proofing by emitting a warning and aborting the UserSaveOptions
hook handler if an empty originalOptions array is received. A stack
trace will also be logged in this case to help track down the
source of the problem.

Bug: T218835
Change-Id: Id26b03a40724b62462a17a4beff37752e70d2e6e