Add more guard statements to VectorPrefDiffInstrumentation
In one of the great mysteries of mediawiki, this instrumentation
sometimes sends events with the `final_state` field set to `vector`
(incorrect) instead of `vector1` (correct) or `vector2` (correct).

Because the cause of this is still unknown and doesn't make sense to me
based on the code which should always return either '1' or '2' if
executed [1], I've added code to log this unexpected state to logstash
when it occurs. Additionally, the code will return early to avoid
sending bad data.

Additional changes:

* Change the `in_array` call to use strict comparison

* Add more tests for unexpected state

* Explicitly set `wgWMEVectorPrefDiffSalt` in testNullSalt (I ran into
problems locally with this test because my LocalSettings had set this
variable to not be null).

[1] https://github.com/wikimedia/Vector/blob/402b01e0cae4df050dd24d3e3f99b0272fc12122/includes/HTMLForm/Fields/HTMLLegacySkinVersionField.php#L65-L69

Bug: T261842
Change-Id: I928fc70b94e3bc7cde1fa72e68499cfd730d4f94