Fix 'final_state: vector' bug in VectorPrefDiffInstrumentation
If the following conditions are met, this instrumentation would send an
event with the `final_state` field set to `vector` (incorrect) instead
of `vector1` (correct) or 'vector2' (correct):

* A user uses GlobalPreferences form and saves the GlobalPreferences
`VectorSkinVersion` form field **unchecked** (corresponding to modern
Vector).

* The above user then changes to the Vector skin from another skin (e.g.
from Minerva to Vector ) using the default Preferences form (not the
GlobalPreferences form) and **without** having checked the "Set a local
exception for this global preference" for the `VectorSkinVersion` field.

This occurs because:

* The GlobalPreferences extension disables any form field that the user
has set with the GlobalPreferences form if the user is on the default
preferences form (not using the GlobalPreferences form) AND if there is
not a "local exception" for the form field. [1]

* If a form field is disabled, the `HTMLForm` class calls the field's
`getDefault` method instead of the field's `loadDataFromRequest` method.
The former call returns a bool while the latter call returns a string in
the form of '1' or '2'. [2]

* Not knowing the preceding point, this instrumentation expected the
`VectorSkinVersion` field to always return a string of either `1` or `2`
from the field's `loadDataFromRequest` method [3]. Regardless of whether
the `VectorSkinVersion` form field returned a `bool` or a `string`, this
instrumentation would try to concat the value with the skin name. [4] In
PHP, `"vector" . false === "vector"` (which reproduces the issue). What
make this even more confusing is that `"vector" . true === "vector1"`
(which doesn't reproduce the issue).

To fix this instrumentation's handling of such data, the
`generateSkinVersionName` method was refactored to correctly handle a
skin version that is either a bool or a string.

[1] https://github.com/wikimedia/mediawiki-extensions-GlobalPreferences/blob/ccf4c9d470bfc0714119153b592bcc167dceccc6/includes/GlobalPreferencesFactory.php#L175
[2] https://github.com/wikimedia/mediawiki/blob/fca9c972de9333bfbf881fdaa639abe27b9de4da/includes/htmlform/HTMLForm.php#L1861-L1865
[3] https://github.com/wikimedia/Vector/blob/25ee08077345e0709fc9185ced9dd8757653b4ae/includes/HTMLForm/Fields/HTMLLegacySkinVersionField.php#L65-L69
[4] https://github.com/wikimedia/mediawiki-extensions-WikimediaEvents/blob/92450158484308e2416a2591bd7e2dcf278f3cbd/includes/VectorPrefDiffInstrumentation.php#L92

Bug: T261842
Change-Id: I668d86fe752afd6766c763bc4c39aa7464450792
(cherry picked from commit cb6a48de40125cb0fc624c0ad67a61816e878944)