Use method provided by the BetaFeatures extension to check changes of beta preferences
Other extensions shouldn't rely on the value of user options of beta features, which are managed by the BetaFeatures extension.
It works fine till now, but the auto-enroll logic of BetaFeatures extension needs a overhaul and transit to soft enroll, which means the value can't reflect the real enable state of features.

Depends-On: I2ba8c65266db8d12300e36307324b4dfe82c4028
Change-Id: I8ca3d265bdb68abfd12f0b93cd71710ceff4756b