Add rich snippet instrument
- Add page split testing for bucketing sampled pages randomly in service
of an experiment.
- Add hashing of the page id to randomly and consistently bucket
sampled pages (treatment + control).
- Add new hook to manipulate the robots meta tag for appending the
max-snippet directive for pages in the treatment group.
- Inside the hook "onXAnalyticsSetHeader", update the X-Analytics header
with a new key-value pair called "max-snippet" (boolean) to determine
if pages in the control and treatment groups have the rich snippet.
Value is false for the control group, true for the treatment group.
Key-value pair is not added to unsampled pages.
- Add unit tests for the new classes.

Note: This instrument was adapted from the PageSplitTester functionality
that was introduced in T206868 in the Wikibase extension. Please see the
original description of the relevant patch at
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Wikibase/+/469926
for more details about the bucketing and sampling of pages, configuration.

Because of performance degradation concerns using the PageRandomLookup
class from the original effort, the bucketing mechanism introduced in
this patch uses a hashed page id to evenly distribute sampled pages
between treatment and control group buckets.

Bug: T301584
Depends-On: I2652f20c9fe5f2d3e28e329d4eddc2a2ff88cf76
Change-Id: I4ae2c0b1cdc71ea6631bd88544dc360293c40af1