Set special=unknown on X-Analytics
When the special page name does not resolve, if we don't set special in
the X-Analytics header, downstream consumers may assume this is not a
special page.  This caused an issue in the pageview definition which
relied on special being set.  We considered filtering out special pages
in some other way, but it would add lots of unnecessary string
processing to an operation that's run many tens of thousands of times
per second.  Instead, this solution is more efficient and closer to the
original intent of the code as well.

Bug: T304362
Change-Id: Iddecbc7d834d30c84d1a27c8f18679324e63ae1b