*WebUIActionsTracking: Also log events via Metrics Platform
The first step of migrating the Desktop- and MobileWebUIActionsTracking
instruments to the Metrics Platform [0] is to log events using both the
Event Platform (albeit indireclty) and the Metrics Platform client
directly.

Note well that the Metrics Platform client will not log an event without
one or more streams being configured to receive that event. Therefore,
this change is a NOP.

Additionally, the Metrics Platform client can mix in additional
information - so-called "contextual attributes" [1] - based on the
stream configuration. The majority of the values in the *WebUIActions
schemas can also be mixed in by the Metrics Platform client.

An example stream configuration is given in [2].

[0] https://wikitech.wikimedia.org/wiki/Metrics_Platform
[1] https://gerrit.wikimedia.org/g/mediawiki/libs/metrics-platform/+/aed6738b845/js/src/StreamConfig.d.ts#31
[2] https://phabricator.wikimedia.org/P30427

Bug: T311268
Change-Id: I9385d615ee3e610ba3407085acccd3d41ac3504f