Remove unused support for `mw.track( 'wikimedia.event.*' )`
As far as I can tell, nothing uses this:
https://codesearch.wmcloud.org/search/?q=%5Cbwikimedia%5C.event%5Cb

It was added for T138659, but even in that task the feature was never
used (as noted in the last comment there).

The original problem that led to this being added should be resolved
by the work related to T205744 (non-existent schemas no longer cause
any visible issues).

Change-Id: I28c925cfda4b003c376b45e0fddde41da6c5efe5