sessionTick: Log localStorage support
If localStorage is not supported - for whatever reason - then log it as
part of the event. This will allow us to provide an estimate for data
loss if the instrument were updated to use localStorage.

Note well that we only log if localStorage is not supported. This should
minimize bytes transferred to the event ingestion servers, provided that
the majority of browsers support localStorage.

Other changes:

* Remove the test.qa temporary test value, introduced as part of
  T276515 and flagged for removal in T276515#6914300

Bug: T295619
Bug: T276515
Change-Id: Idb5b3f85a3561e84e9ed5c28f09a34b9e92e0305