Distinguish between pinned/unpinned elements in desktopWebUIActions
With the ability to 'pin' menus to the sidebar in Vector 2022,
we have to distinguish between clicks that occur on pinned vs
unpinned links (which have the same ID in both cases).

Now before sending the click event, we search the parent
element for a pinned header, and add the pinned/unpinned state,
as well as the name of that feature to the instrumentation data.
e.g. instead of `ca-protect` now `ca-protect.vector-page-tools.pinned`
as the el_id value.

This change also unifies the two click handlers (see related discusson
on I5ef98d5f5713d3d99bc5f7f8112ba2d1e0f62e22) to avoid logging duplicate
events for the same link given the new added complexity to this logic.

Bug: T324878
Change-Id: I3c3f0215d2f8b638a4ff9a2f0ca97cdc0e5b69f1